Fix coding standard problems
