+++ tests/ZendTest/Db/Sql/Ddl/Column/AbstractLengthColumnTest.php
