+++ tests/ZendTest/Db/Sql/Ddl/Column/AbstractPrecisionColumnTest.php
